fix: gracefully skip ViT when server memory insufficient
- Revert fp16 (not supported on CPU)
- ViT load failure catches exception, falls to CNN+AI dual branch
- Fusion weights: CNN 55%, ViT 25%, AI 20% (adapted for 4GB server)

+++ backend/app/detectors/image/vit_branch.py
@@ -58,7 +58,7 @@ class ViTBranch(DetectionPipeline):
             local_cache = "../models/image/clip-vit-large-patch14"
             if _os.path.isdir(local_cache) and _os.path.exists(_os.path.join(local_cache, "config.json")):
                 model_path = local_cache
-            self._model = CLIPModel.from_pretrained(model_path, local_files_only=True, torch_dtype=torch.float16).vision_model.to(self.device)
+            self._model = CLIPModel.from_pretrained(model_path, local_files_only=True).vision_model.to(self.device)
             try:
                 self._processor = CLIPImageProcessor.from_pretrained(model_path, local_files_only=True)
             except Exception:
